requirements.txt
Django>=4.2,<5.0
requests>=2.31.0
django-allauth>=0.57.0
PyJWT>=2.10.0
cryptography>=46.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
